What's this about: Technically undo previous commit.
Add functionality to view QR of a Pager, not that this code is a result
of a huge slicing, due to no integration to other function for now, had
to use placeholder for neccesary part of the code.

Huge update: Recreate and modify Dart files and directory straight from main branch to avoid
'git pull' drama, rename the existing code base and integrate it with
the new Dart files, and add new dependency inside pubspec.yaml for
screen utility and Pretty QR view for QR payment.

It should works now, turns out the route for QRDetailView class hasn't
been registered in main file.